+++ doc/user/content/release-notes.md
@@ -48,13 +48,23 @@ Wrap your release notes at the 80 character mark.
 
 {{% version-header v0.7.2 %}}
 
+- Introduce the concept of [volatility](/overview/volatility) to describe
+  sources that do not provide reliability guarantees that
+  Materialize relies on. The new volatility information is surfaced via
+  [`SHOW SOURCES`](/sql/show-sources), [`SHOW VIEWS`](/sql/show-views),
+  and [`SHOW SINKS`](/sql/show-sinks).
+
+- Add [PubNub sources](/sql/create-source/text-pubnub).
+
+- Add [`S3` sources](/sql/create-source/text-s3).
+
 - Add a [`--log-filter` command-line option](/cli/#logging) and a
   `MZ_LOG_FILTER` environment variable that control which log messages to emit.
 
   This behavior was previously available via the undocumented `MZ_LOG`
   environment variable, which will be removed in a future release.
 
-- Record Kafka Consumer metrics in the `mz_kafka_consumer_statistics` system
+- Record Kafka Consumer metrics in the `mz_kafka_consumer_partitions` system
   table. Enabled by default for all Kafka sources.
 
 - Add the [`jsonb_object_agg`](/sql/functions/jsonb_object_agg) function to